src/backend/requirements/prod.txt
python-terraform==0.10.1
sqlalchemy==1.3.18
fastapi==0.61.0
uvicorn==0.11.8
celery[redis,msgpack]==4.4.7
gunicorn==20.0.4
psycopg2-binary==2.8.5
alembic==1.4.2
paramiko==2.7.1
pillow==7.2.0
pytesseract==0.3.5
terrascript==0.8.0
supervisor==4.2.0
vultr==1.0.1
msgpack-python==0.5.6
python-dotenv==0.14.0
fabric==2.5.0
python-digitalocean==1.15.0
python-namesilo==1.1.3
pyjwt==1.7.1
passlib[bcrypt]==1.7.2
beautifulsoup4==4.9.1
lxml==4.5.2
python-multipart==0.0.5
tencentcloud-sdk-python==3.0.235
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-ecs==4.19.10